refactor: move kit to `@nuxt/devtools-kit`
