+++ docs/content/2.module/0.guide.md
@@ -23,7 +23,7 @@ Currently the only way to contribute to Nuxt DevTools View is via iframe. You ne
 You can use the utility kit provided by Nuxt DevTools to register your custom tab:
 
 ```ts
-import { addCustomTab } from '@nuxt/devtools/kit'
+import { addCustomTab } from '@nuxt/devtools-kit'
 
 addCustomTab({
   // unique identifier
