+++ docs/content/2.module/1.utils-kit.md
@@ -6,20 +6,20 @@ APIs are subject to change.
 
 Since v0.3.0, we are now providing a utility kit for easier DevTools integrations, similar to `@nuxt/kit`.
 
-It can be access via `@nuxt/devtools/kit`:
+It can be access via `@nuxt/devtools-kit`:
 
 ```ts
-import { addCustomTab } from '@nuxt/devtools/kit'
+import { addCustomTab } from '@nuxt/devtools-kit'
 ```
 
-Generally, we recommend to module authors to install `@nuxt/devtools` as a dev dependency and bundled `@nuxt/devtools/kit` into your module.
+Generally, we recommend to module authors to install `@nuxt/devtools` as a dev dependency and bundled `@nuxt/devtools-kit` into your module.
 
 ## `addCustomTab()`
 
 A shorthand for calling the hook `devtools:customTabs`.
 
 ```ts
-import { addCustomTab } from '@nuxt/devtools/kit'
+import { addCustomTab } from '@nuxt/devtools-kit'
 
 addCustomTab(() => ({
   // unique identifier
@@ -45,7 +45,7 @@ A shorthand for call hook `devtools:customTabs:refresh`. It will refresh all cus
 Start a sub process using `execa` and create a terminal tab in DevTools.
 
 ```ts
-import { startSubprocess } from '@nuxt/devtools/kit'
+import { startSubprocess } from '@nuxt/devtools-kit'
 
 const subprocess = startSubprocess(
   {
