+++ packages/devtools-kit/build.config.ts
@@ -1,0 +1,12 @@
+import { defineBuildConfig } from 'unbuild'
+
+export default defineBuildConfig({
+  externals: [
+    'nuxt',
+    'nuxt/schema',
+    'vite',
+    'vue',
+    '@nuxt/kit',
+    '@nuxt/schema',
+  ],
+})
