+++ packages/devtools-kit/package.json
@@ -1,0 +1,39 @@
+{
+  "name": "@nuxt/devtools-kit",
+  "type": "module",
+  "version": "0.2.5",
+  "license": "MIT",
+  "repository": "nuxt/devtools",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "require": "./dist/index.cjs",
+      "import": "./dist/index.mjs"
+    },
+    "./types": {
+      "types": "./dist/types.d.ts",
+      "require": "./dist/types.cjs",
+      "import": "./dist/types.mjs"
+    }
+  },
+  "main": "./dist/index.cjs",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist",
+    "*.d.ts",
+    "*.cjs",
+    "*.mjs"
+  ],
+  "scripts": {
+    "build": "unbuild",
+    "prepublishOnly": "pnpm build"
+  },
+  "dependencies": {
+    "@nuxt/kit": "^3.2.3",
+    "@nuxt/schema": "^3.2.3",
+    "execa": "^7.1.0"
+  },
+  "devDependencies": {
+    "unbuild": "^1.1.2"
+  }
+}
