+++ packages/devtools-kit/src/types/custom-tabs.ts
