+++ packages/devtools-kit/src/types/index.ts
@@ -1,0 +1,4 @@
+import './hooks'
+
+export * from './custom-tabs'
+export * from './terminals'
