+++ packages/devtools/client/components/SideNav.vue
@@ -4,7 +4,7 @@ const tabs = useTabs()
 </script>
 
 <template>
-  <div border="r base" flex="~ col gap-0.5" h-full items-center z-100 bg-base>
+  <div border="r base" flex="~ col gap-0.5" h-full items-center bg-base z-100>
     <div flex="~ col" items-center bg-base top-0 pt3 sticky z-1 mb1>
       <VDropdown placement="left-start" :distance="20">
         <button
