+++ packages/devtools/package.json
@@ -33,9 +33,6 @@
     "*.cjs",
     "*.mjs"
   ],
-  "workspaces": [
-    "client"
-  ],
   "scripts": {
     "build": "pnpm dev:prepare && pnpm build:module && pnpm build:client && cp ../../README.md .",
     "build:client": "nuxi generate client",
@@ -51,6 +48,7 @@
   },
   "dependencies": {
     "@antfu/install-pkg": "^0.1.1",
+    "@nuxt/devtools-kit": "workspace:*",
     "@nuxt/kit": "^3.2.3",
     "birpc": "^0.2.7",
     "consola": "^2.15.3",
