+++ packages/devtools/src/kit.ts
@@ -1,1 +1,1 @@
-export * from './kit/index'
+export * from '@nuxt/devtools-kit'
