+++ packages/devtools/src/module.ts
@@ -1,7 +1,6 @@
 import { defineNuxtModule } from '@nuxt/kit'
 import { defaultOptions } from './constant'
 import type { ModuleGlobalOptions, ModuleOptions } from './types'
-import './types/hooks'
 import { isGlobalInstall } from './dirs'
 
 export type { ModuleOptions }
