+++ packages/devtools/src/types/index.ts
@@ -1,6 +1,7 @@
+export * from '@nuxt/devtools-kit/types'
+
 export * from './integrations'
 export * from './module'
 export * from './rpc'
 export * from './ui-state'
 export * from './client-api'
-export * from './custom-tabs'
