+++ tsconfig.json
@@ -12,6 +12,8 @@
     "skipLibCheck": true,
     "paths": {
       "@nuxt/devtools": ["./packages/devtools/src/index.ts"],
+      "@nuxt/devtools-kit": ["./packages/devtools-kit/src/index.ts"],
+      "@nuxt/devtools-kit/types": ["./packages/devtools-kit/src/types/index.ts"],
       "@nuxt/devtools/iframe-client": ["./packages/devtools/src/iframe-client.ts"]
     }
   },
